requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.2
    # via -r requirements.in
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
billiard==4.2.0
    # via celery
black==24.8.0
    # via -r requirements.in
celery==5.4.0
    # via -r requirements.in
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
fastapi==0.112.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via -r requirements.in
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
kombu==5.4.0
    # via celery
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via mako
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.47
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
pytest==8.3.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.0.1
    # via -r requirements.in
redis==5.0.8
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.31
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via celery
urllib3==2.2.2
    # via requests
uuid7==0.1.0
    # via -r requirements.in
uvicorn==0.30.5
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
